+++ src/screens/SelectExerciseScreen.js
@@ -1,6 +1,5 @@
 import React, { Component } from "react";
 import { ScrollView, View, Fragment } from "react-native";
-
 import {
   Text,
   Button,
@@ -14,6 +13,8 @@ import {
   CheckBox,
   Content
 } from "native-base";
+
+import SelectList from "../components/SelectList";
 
 export default class SelectExerciseScreen extends Component {
   constructor(props) {
@@ -40,41 +41,11 @@ export default class SelectExerciseScreen extends Component {
     this.props.callback_setSelectedMuscleGroup(selectedMuscleGroup);
   };
 
-  toggleSelected = selected => {
-    const indexOf = this.state.selected.indexOf(selected);
-
-    if (indexOf >= 0) {
-      let newSelectedArray = this.state.selected;
-      newSelectedArray.splice(indexOf, 1);
-      this.setState(prevState => ({
-        selected: [...newSelectedArray]
-      }));
-    } else {
-      this.setState(prevState => ({
-        selected: [...prevState.selected, selected]
-      }));
-    }
-  };
-
-  renderSelectList = list => {
-    return list.map(item => {
-      return (
-        <View key={item} style={{ padding: 4, width: "100%" }}>
-          <Button
-            full
-            warning
-            bordered={!this.state.selected.includes(item)}
-            rounded
-            style={{ height: 40 }}
-            onPress={() => {
-              this.toggleSelected(item);
-            }}
-          >
-            <Text style={{ color: "#000" }}>{item}</Text>
-          </Button>
-        </View>
-      );
-    });
+  setSelected = selectedList => {
+    alert(selectedList);
+    this.setState(() => ({
+      selected: [...selectedList]
+    }));
   };
 
   handleBuildWorkoutPress = () => {
@@ -104,8 +75,12 @@ export default class SelectExerciseScreen extends Component {
           >
             <Text style={{ paddingTop: 8, paddingBottom: 8 }}>
               Select which muscle groups you want to add to your workout!
+              {JSON.stringify(this.state)}
             </Text>
-            {this.renderSelectList(this.state.muscleGroups)}
+            <SelectList
+              list={this.state.muscleGroups}
+              callback_setSelected={this.setSelected}
+            />
             <Button
               block
               primary
